requirements.txt
anthropic>=0.25.0
google-generativeai>=0.7.0
moviepy>=1.0.3
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
